+++ src/app/model/specs/test-utils.tsx
@@ -1,9 +1,9 @@
-import { createRouter, RouterProvider } from '@tanstack/react-router';
+import { type AnyRouter, RouterProvider } from '@tanstack/react-router';
 import { render } from '@testing-library/react';
 
 import { ConfirmDialogProvider } from '@shared';
 
-export const renderApp = (router: ReturnType<typeof createRouter>): ReturnType<typeof render> =>
+export const renderApp = (router: AnyRouter): ReturnType<typeof render> =>
   render(
     <ConfirmDialogProvider>
       <RouterProvider router={router} />