+++ src/shared/model/helpers/dates.ts
@@ -1,7 +1,9 @@
 import { Nullable } from '@alexevs/ts-guards';
 
 export const formatDate = (date: Nullable<string>): string => {
-  if (!date) return 'Never';
+  if (!date) {
+    return 'Never';
+  }
   return new Date(date).toLocaleDateString('en-US', {
     day: 'numeric',
     month: 'short',