+++ RELEASE-NOTES.md
@@ -1,9 +1,18 @@
+# Release v0.2.0
+## What's Changed
+* [Common] RELEASE-NOTES.md update v0.1.2 by @a1exevs in https://github.com/a1exevs/set-forge/pull/6
+* [Refactoring] Consolidating Cursor Commands and Adding Component Specs by @a1exevs in https://github.com/a1exevs/set-forge/pull/7
+* [Feature] Adding shared Dialog component by @a1exevs in https://github.com/a1exevs/set-forge/pull/8
+* [Feature] Adding Confirm Dialog Component by @a1exevs in https://github.com/a1exevs/set-forge/pull/9
+* [Documentation][Bugfix] Adding specs and fixing edge cases by @a1exevs in https://github.com/a1exevs/set-forge/pull/10
+* [Feature] Workout list editing: Edit page, WorkoutListForm widget, and MenuButton by @a1exevs in https://github.com/a1exevs/set-forge/pull/11
+* [Common] Version increase v0.2.0 by @a1exevs in https://github.com/a1exevs/set-forge/pull/12
+* [Bugfix] Story runtime error by @a1exevs in https://github.com/a1exevs/set-forge/pull/14
+
+**Full Changelog**: https://github.com/a1exevs/set-forge/compare/v0.1.2...v0.2.0
+
 # Release v0.1.2
 ## What's Changed
-* [Feature] Initial Project Setup with FSD Architecture by @a1exevs in https://github.com/a1exevs/set-forge/pull/1
-* [Common][Feature][Tests][Storybook] MVP by @a1exevs in https://github.com/a1exevs/set-forge/pull/3
-
-**Full Changelog**: https://github.com/a1exevs/set-forge/commits/v0.1.2## What's Changed
 * [Feature] Initial Project Setup with FSD Architecture by @a1exevs in https://github.com/a1exevs/set-forge/pull/1
 * [Common][Feature][Tests][Storybook] MVP by @a1exevs in https://github.com/a1exevs/set-forge/pull/3